Revert "Ver editar campo"
