+++ package.json
@@ -4,7 +4,6 @@
   "private": true,
   "dependencies": {
     "@mapbox/mapbox-gl-draw": "^1.4.2",
-    "@mapbox/mapbox-gl-draw-static-mode": "^1.0.1",
     "@mapbox/mapbox-gl-geocoder": "^5.0.1",
     "@testing-library/jest-dom": "^5.16.5",
     "@testing-library/react": "^13.4.0",
@@ -13,19 +12,17 @@
     "@types/mapbox-gl": "^2.7.11",
     "@urbica/react-map-gl-draw": "^0.3.5",
     "axios": "^1.4.0",
-    "bootstrap": "^5.3.1",
+    "bootstrap": "^5.3.0",
     "bootstrap-icons": "^1.10.5",
     "chart.js": "^4.3.2",
     "cors": "^2.8.5",
     "date-fns": "^2.30.0",
     "dotenv": "^16.3.1",
-    "lodash": "^4.17.21",
     "mapbox-gl": "^2.15.0",
     "mapbox-gl-draw-waypoint": "^1.0.3",
     "papaparse": "^5.4.1",
     "prop-types": "^15.8.1",
     "react": "^18.2.0",
-    "react-bootstrap": "^2.8.0",
     "react-burger-menu": "^3.0.9",
     "react-dom": "^18.2.0",
     "react-drag-drop-files": "^2.3.10",
