+++ src/App.css
@@ -2,10 +2,37 @@
   text-align: center;
 }
 
-/* @media screen and (max-width:80em) {
-  .App{
-    width: 100rem;
-    margin: 0;
-    height: 130rem;
+.App-logo {
+  height: 40vmin;
+  pointer-events: none;
 }
-} */
+
+@media (prefers-reduced-motion: no-preference) {
+  .App-logo {
+    animation: App-logo-spin infinite 20s linear;
+  }
+}
+
+.App-header {
+  background-color: #282c34;
+  min-height: 100vh;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  font-size: calc(10px + 2vmin);
+  color: white;
+}
+
+.App-link {
+  color: #61dafb;
+}
+
+@keyframes App-logo-spin {
+  from {
+    transform: rotate(0deg);
+  }
+  to {
+    transform: rotate(360deg);
+  }
+}
