+++ src/App.js
@@ -2,9 +2,8 @@ import './App.css';
 import { Route, Routes } from 'react-router-dom';
 import React from 'react';
 import AgregarCampo from './componentes/agregarCampo/agregarCampo';
-import EditarCampo from './componentes/editarCampo/editarCampo';
-import VerCampo from './componentes/verCampo/verCampo';
 import AgregarLotes from './componentes/agregarLotes/agregarLotes';
+import EditarCampo from './componentes/editarCampo/editarCampo';
 import Home from './componentes/home/home';
 import IniciarSesion from './componentes/iniciarSesion/iniciarSesion';
 import Registrarse from './componentes/registrarse/registrarse';
@@ -20,13 +19,12 @@ import PswUpdated from './componentes/pswUpdated/pswUpdated';
 function App() {
   document.title = 'AGROIA';
   return (
-    <div className="App" id="App">
+    <div className="App">
       <Routes>
         <Route path="/" element={<HomePrincipal />} />
         <Route path="/agregarCampo/:userID" element={<AgregarCampo />} />
         <Route path="/agregarLotes/:userID" element={<AgregarLotes />} />
-        <Route path="/editarCampo/:userID/:fieldID" element={<EditarCampo />} />
-        <Route path="/verCampo/:userID/:fieldID" element={<VerCampo />} />
+        <Route path="/editarCampo/:userID" element={<EditarCampo />} />
         <Route path="/home/:userID" element={<Home />} />
         <Route path="/iniciarSesion" element={<IniciarSesion />} />
         <Route path="/registrarse" element={<Registrarse />} />
