+++ src/componentes/reusable/boton/button.scss
@@ -12,11 +12,6 @@
     border: solid;
     border-style: 1px $main-green; 
     font-family: "Inter", sans-serif;
-    
-  }
-
-  .mas-button{
-    width: 90%;
   }
 
   .green-button:hover{
