+++ src/componentes/reusable/header/header.scss
@@ -4,8 +4,6 @@
     width: 100%;
     margin-top: 0%;
     @include alignHorizontally;
-    // position: sticky !important;
-    display: block;
 }
