--- src/componentes/reusable/loader/loader.js
@@ -1,11 +1,0 @@
-import React from 'react';
-import './loader.scss';
-import Icon from '../../../assets/icons/icon';
-
-export default function Loader() {
-  return (
-    <div className="loader2">
-      <Icon className="bi bi-flower1 flower-loader" color="#439775" fontSize="7rem" />
-    </div>
-  );
-}
