--- src/componentes/reusable/loader/loader.scss
@@ -1,31 +1,0 @@
-@import "../../../assets/global.scss";
-
-.loader {
-  margin: auto;
-  border: 20px solid #EAF0F6;
-  border-radius: 50%;
-  border-top: 20px solid $verde-principal;
-  width: 5rem;
-  height: 5rem;
-  animation: spinner 2s linear infinite;
-  align-items: center;
-}
-
-.loader2 {
-  margin: auto;
-  border: 20px solid transparent;
-  border-radius: 50%;
-  border-top: 20px solid transparent;
-  width: 10rem;
-  height: 10rem;
-  animation: spinner 3s linear infinite;
-}
-
-.flower-loader{
-animation: spinner 2s linear infinite;
-}
-
-@keyframes spinner {
-  0% { transform: rotate(0deg); }
-  100% { transform: rotate(360deg); }
-}
