--- src/componentes/reusable/map/campoPrueba.js
@@ -1,247 +1,0 @@
-/* eslint-disable import/prefer-default-export */
-export const campoPrueba = {
-  field: {
-    name: 'gez',
-    coordinates: { lat: -34.68624841398288, lon: -59.03278391090079 },
-    plots: [
-      {
-        crop: 'none',
-      },
-      {
-        crop: 'none',
-      },
-      {
-        crop: 'none',
-      },
-      {
-        crop: 'none',
-      },
-      {
-        crop: 'none',
-      },
-      {
-        crop: 'none',
-      },
-      {
-        crop: 'sunflower',
-      },
-      {
-        crop: 'sunflower',
-      },
-      {
-        crop: 'sunflower',
-      },
-      {
-        crop: 'sunflower',
-      },
-      {
-        crop: 'none',
-      },
-      {
-        crop: 'none',
-      },
-      {
-        crop: 'none',
-      },
-      {
-        crop: 'none',
-      },
-      {
-        crop: 'sunflower',
-      },
-      {
-        crop: 'sunflower',
-      },
-      {
-        crop: 'sunflower',
-      },
-      {
-        crop: 'sunflower',
-      },
-      {
-        crop: 'sunflower',
-      },
-      {
-        crop: 'sunflower',
-      },
-      {
-        crop: 'sunflower',
-      },
-      {
-        crop: 'none',
-      },
-      {
-        crop: 'soy',
-      },
-      {
-        crop: 'soy',
-      },
-      {
-        crop: 'soy',
-      },
-      {
-        crop: 'sunflower',
-      },
-      {
-        crop: 'sunflower',
-      },
-      {
-        crop: 'sunflower',
-      },
-      {
-        crop: 'sunflower',
-      },
-      {
-        crop: 'sunflower',
-      },
-      {
-        crop: 'sunflower',
-      },
-      {
-        crop: 'sunflower',
-      },
-      {
-        crop: 'sunflower',
-      },
-      {
-        crop: 'soy',
-      },
-      {
-        crop: 'soy',
-      },
-      {
-        crop: 'soy',
-      },
-      {
-        crop: 'sunflower',
-      },
-      {
-        crop: 'sunflower',
-      },
-      {
-        crop: 'sunflower',
-      },
-      {
-        crop: 'sunflower',
-      },
-      {
-        crop: 'sunflower',
-      },
-      {
-        crop: 'sunflower',
-      },
-      {
-        crop: 'sunflower',
-      },
-      {
-        crop: 'sunflower',
-      },
-      {
-        crop: 'soy',
-      },
-      {
-        crop: 'soy',
-      },
-      {
-        crop: 'soy',
-      },
-      {
-        crop: 'soy',
-      },
-      {
-        crop: 'sunflower',
-      },
-      {
-        crop: 'sunflower',
-      },
-      {
-        crop: 'sunflower',
-      },
-      {
-        crop: 'sunflower',
-      },
-      {
-        crop: 'sunflower',
-      },
-      {
-        crop: 'sunflower',
-      },
-      {
-        crop: 'none',
-      },
-      {
-        crop: 'soy',
-      },
-      {
-        crop: 'soy',
-      },
-      {
-        crop: 'soy',
-      },
-      {
-        crop: 'soy',
-      },
-      {
-        crop: 'sunflower',
-      },
-      {
-        crop: 'sunflower',
-      },
-      {
-        crop: 'sunflower',
-      },
-      {
-        crop: 'sunflower',
-      },
-      {
-        crop: 'none',
-      },
-      {
-        crop: 'none',
-      },
-      {
-        crop: 'none',
-      },
-      {
-        crop: 'none',
-      },
-      {
-        crop: 'soy',
-      },
-      {
-        crop: 'soy',
-      },
-      {
-        crop: 'soy',
-      },
-      {
-        crop: 'sunflower',
-      },
-      {
-        crop: 'none',
-      },
-      {
-        crop: 'none',
-      },
-      {
-        crop: 'none',
-      },
-      {
-        crop: 'none',
-      },
-      {
-        crop: 'none',
-      },
-      {
-        crop: 'none',
-      },
-    ],
-    height: 7,
-    width: 11,
-    image: 'image.png',
-    _id: '64cc0434fec4089b438d29ea',
-    createdAt: '2023-08-03T19:47:00.514Z',
-    updatedAt: '2023-08-03T19:47:00.514Z',
-    __v: 0,
-  },
-};
