+++ src/componentes/reusable/map/styles.js
@@ -25,7 +25,7 @@ const styles = [
     ],
     paint: {
       'fill-color': '#fbb03b',
-      'fill-outline-color': '#0d0d0d',
+      'fill-outline-color': '#fbb03b',
       'fill-opacity': 0.1,
     },
   },
@@ -52,7 +52,7 @@ const styles = [
       'line-join': 'round',
     },
     paint: {
-      'line-color': '#0d0d0d',
+      'line-color': '#3bb2d0',
       'line-width': 2,
     },
   },
@@ -253,33 +253,6 @@ const styles = [
       'fill-color': ['get', 'user_portColor'],
       'fill-outline-color': ['get', 'user_portColor'],
       'fill-opacity': 0.5,
-    },
-  },
-  { // ver campo
-    id: 'ver-campo-ndvi',
-    type: 'fill',
-    filter: ['all', ['==', '$type', 'Polygon'],
-      ['has', 'user_fillColor'],
-    ],
-    paint: {
-      'fill-color': ['get', 'user_fillColor'],
-      'fill-outline-color': 'transparent',
-      'fill-opacity': 0.2,
-    },
-  },
-  { // ver campo
-    id: 'ver-campo-ndvi-lines',
-    type: 'line',
-    filter: ['all', ['==', '$type', 'Polygon'],
-      ['has', 'user_fillColor'],
-    ],
-    layout: {
-      'line-cap': 'round',
-      'line-join': 'round',
-    },
-    paint: {
-      'line-width': 0,
-      'line-opacity': 0,
     },
   },
   {
